Style: CategoryService 통합
